[INLONG-5296][Manager] Add fields for DataNode entity
