+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/cluster/InlongClusterServiceImpl.java
@@ -39,19 +39,6 @@ import org.apache.inlong.manager.common.enums.GroupStatus;
 import org.apache.inlong.manager.common.enums.MQType;
 import org.apache.inlong.manager.common.enums.UserTypeEnum;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
-import org.apache.inlong.manager.pojo.cluster.BindTagRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterInfo;
-import org.apache.inlong.manager.pojo.cluster.ClusterNodeRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterNodeResponse;
-import org.apache.inlong.manager.pojo.cluster.ClusterPageRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterTagPageRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterTagRequest;
-import org.apache.inlong.manager.pojo.cluster.ClusterTagResponse;
-import org.apache.inlong.manager.pojo.cluster.pulsar.PulsarClusterDTO;
-import org.apache.inlong.manager.pojo.group.InlongGroupBriefInfo;
-import org.apache.inlong.manager.pojo.group.InlongGroupPageRequest;
-import org.apache.inlong.manager.pojo.stream.InlongStreamBriefInfo;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.common.util.Preconditions;
 import org.apache.inlong.manager.dao.entity.InlongClusterEntity;
@@ -65,6 +52,19 @@ import org.apache.inlong.manager.dao.mapper.InlongClusterTagEntityMapper;
 import org.apache.inlong.manager.dao.mapper.InlongGroupEntityMapper;
 import org.apache.inlong.manager.dao.mapper.InlongStreamEntityMapper;
 import org.apache.inlong.manager.dao.mapper.UserEntityMapper;
+import org.apache.inlong.manager.pojo.cluster.BindTagRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterInfo;
+import org.apache.inlong.manager.pojo.cluster.ClusterNodeRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterNodeResponse;
+import org.apache.inlong.manager.pojo.cluster.ClusterPageRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterTagPageRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterTagRequest;
+import org.apache.inlong.manager.pojo.cluster.ClusterTagResponse;
+import org.apache.inlong.manager.pojo.cluster.pulsar.PulsarClusterDTO;
+import org.apache.inlong.manager.pojo.group.InlongGroupBriefInfo;
+import org.apache.inlong.manager.pojo.group.InlongGroupPageRequest;
+import org.apache.inlong.manager.pojo.stream.InlongStreamBriefInfo;
 import org.apache.inlong.manager.service.repository.DataProxyConfigRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -215,7 +215,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
                     String updateTags = Joiner.on(",").join(tagSet);
                     entity.setClusterTags(updateTags);
                     entity.setModifier(operator);
-                    int rowCount = clusterMapper.updateByIdSelective(entity);
+                    int rowCount = clusterMapper.updateById(entity);
                     if (rowCount != InlongConstants.AFFECTED_ONE_ROW) {
                         LOGGER.error("cluster has already updated with name={}, type={}, curVersion={}",
                                 entity.getName(), entity.getType(), entity.getVersion());
@@ -405,7 +405,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
                 InlongClusterEntity updateEntity = clusterMapper.selectById(id);
                 updateEntity.setClusterTags(updateTags);
                 updateEntity.setModifier(operator);
-                int rowCount = clusterMapper.updateByIdSelective(updateEntity);
+                int rowCount = clusterMapper.updateById(updateEntity);
                 if (rowCount != InlongConstants.AFFECTED_ONE_ROW) {
                     LOGGER.error("cluster has already updated with name={}, type={}, curVersion={}",
                             updateEntity.getName(), updateEntity.getType(), updateEntity.getVersion());
@@ -624,7 +624,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
             throw new BusinessException(msg);
         }
 
-        List<InlongClusterEntity> clusterList = clusterMapper.selectByKey(clusterTag, null, ClusterType.DATA_PROXY);
+        List<InlongClusterEntity> clusterList = clusterMapper.selectByKey(clusterTag, null, ClusterType.DATAPROXY);
         if (CollectionUtils.isEmpty(clusterList)) {
             String msg = "not found any data proxy cluster for inlongGroupId=" + inlongGroupId
                     + " and clusterTag=" + clusterTag;
@@ -664,7 +664,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
         ClusterPageRequest request = ClusterPageRequest.builder()
                 .clusterTag(clusterTag)
                 .name(clusterName)
-                .type(ClusterType.DATA_PROXY)
+                .type(ClusterType.DATAPROXY)
                 .build();
         List<InlongClusterEntity> clusterEntityList = clusterMapper.selectByCondition(request);
         DataProxyConfig result = new DataProxyConfig();
@@ -734,7 +734,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
         // get mq cluster info
         LOGGER.debug("GetDPConfig: begin to get mq clusters by tags={}", clusterTagList);
         List<MQClusterInfo> mqSet = new ArrayList<>();
-        List<String> typeList = Arrays.asList(ClusterType.TUBE, ClusterType.PULSAR);
+        List<String> typeList = Arrays.asList(ClusterType.TUBEMQ, ClusterType.PULSAR);
         ClusterPageRequest pageRequest = ClusterPageRequest.builder()
                 .typeList(typeList)
                 .clusterTagList(clusterTagList)
@@ -793,7 +793,7 @@ public class InlongClusterServiceImpl implements InlongClusterService {
         String updateTags = Joiner.on(",").join(tagSet);
         entity.setClusterTags(updateTags);
         entity.setModifier(operator);
-        int rowCount = clusterMapper.updateByIdSelective(entity);
+        int rowCount = clusterMapper.updateById(entity);
         if (rowCount != InlongConstants.AFFECTED_ONE_ROW) {
             LOGGER.error("cluster has already updated with name={}, type={}, curVersion={}", entity.getName(),
                     entity.getType(), entity.getVersion());
