+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/common/utils/ValidationUtilsTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.pojo.common.utils;
+package org.apache.inlong.manager.common.utils;
 
 import org.apache.inlong.manager.pojo.cluster.ClusterTagRequest;
 import org.apache.inlong.manager.common.validation.UpdateValidation;
