+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/service/sink/MySQLStreamSinkServiceTest.java
@@ -15,18 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.core.sink;
+package org.apache.inlong.manager.service.sink;
 
 import org.apache.inlong.manager.common.consts.InlongConstants;
 import org.apache.inlong.manager.common.enums.SinkType;
 import org.apache.inlong.manager.pojo.sink.SinkField;
 import org.apache.inlong.manager.pojo.sink.StreamSink;
-import org.apache.inlong.manager.pojo.sink.hdfs.HDFSSink;
-import org.apache.inlong.manager.pojo.sink.hdfs.HDFSSinkRequest;
+import org.apache.inlong.manager.pojo.sink.mysql.MySQLSink;
+import org.apache.inlong.manager.pojo.sink.mysql.MySQLSinkRequest;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.service.ServiceBaseTest;
 import org.apache.inlong.manager.service.core.impl.InlongStreamServiceTest;
-import org.apache.inlong.manager.service.sink.StreamSinkService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,16 +33,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.util.ArrayList;
 import java.util.List;
 
-public class HDFSStreamSinkServiceTest extends ServiceBaseTest {
+/**
+ * MySQL sink service test
+ */
+public class MySQLStreamSinkServiceTest extends ServiceBaseTest {
 
-    private static final String globalGroupId = "b_group_hdfs";
-    private static final String globalStreamId = "stream1_hdfs";
+    private static final String globalGroupId = "b_group1_mysql";
+    private static final String globalStreamId = "stream1_mysql";
     private static final String globalOperator = "admin";
-    private static final String fileFormat = "TextFile";
-    private static final String dataPath = "hdfs://ip:port/usr/hive/warehouse/test.db";
-    private static final String serverTimeZone = "GMT%2b8";
-    private static final String fieldName = "hdfs_field";
-    private static final String fieldType = "hdfs_type";
+    private static final String fieldName = "mysql_field";
+    private static final String fieldType = "mysql_type";
     private static final Integer fieldId = 1;
 
     @Autowired
@@ -56,24 +55,27 @@ public class HDFSStreamSinkServiceTest extends ServiceBaseTest {
      */
     public Integer saveSink(String sinkName) {
         streamServiceTest.saveInlongStream(globalGroupId, globalStreamId, globalOperator);
+        MySQLSinkRequest sinkInfo = new MySQLSinkRequest();
+        sinkInfo.setInlongGroupId(globalGroupId);
+        sinkInfo.setInlongStreamId(globalStreamId);
+        sinkInfo.setSinkType(SinkType.SINK_MYSQL);
 
-        HDFSSinkRequest hdfsSinkRequest = new HDFSSinkRequest();
-        hdfsSinkRequest.setInlongGroupId(globalGroupId);
-        hdfsSinkRequest.setInlongStreamId(globalStreamId);
-        hdfsSinkRequest.setSinkType(SinkType.SINK_HDFS);
-        hdfsSinkRequest.setEnableCreateResource(InlongConstants.DISABLE_CREATE_RESOURCE);
-        hdfsSinkRequest.setSinkName(sinkName);
-        hdfsSinkRequest.setFileFormat(fileFormat);
-        hdfsSinkRequest.setDataPath(dataPath);
-        hdfsSinkRequest.setServerTimeZone(serverTimeZone);
+        sinkInfo.setJdbcUrl("jdbc:mysql://localhost:5432/database");
+        sinkInfo.setUsername("mysql");
+        sinkInfo.setPassword("inlong");
+        sinkInfo.setTableName("user");
+        sinkInfo.setPrimaryKey("name,age");
+
+        sinkInfo.setSinkName(sinkName);
+        sinkInfo.setEnableCreateResource(InlongConstants.DISABLE_CREATE_RESOURCE);
         SinkField sinkField = new SinkField();
         sinkField.setFieldName(fieldName);
         sinkField.setFieldType(fieldType);
         sinkField.setId(fieldId);
         List<SinkField> sinkFieldList = new ArrayList<>();
         sinkFieldList.add(sinkField);
-        hdfsSinkRequest.setSinkFieldList(sinkFieldList);
-        return sinkService.save(hdfsSinkRequest, globalOperator);
+        sinkInfo.setSinkFieldList(sinkFieldList);
+        return sinkService.save(sinkInfo, globalOperator);
     }
 
     /**
@@ -81,30 +83,29 @@ public class HDFSStreamSinkServiceTest extends ServiceBaseTest {
      */
     public void deleteSink(Integer sinkId) {
         boolean result = sinkService.delete(sinkId, globalOperator);
-        // Verify that the deletion was successful
         Assertions.assertTrue(result);
     }
 
     @Test
     public void testListByIdentifier() {
-        Integer sinkId = this.saveSink("default_hdfs");
-        StreamSink sink = sinkService.get(sinkId);
-        // verify globalGroupId
+        Integer mysqlSinkId = this.saveSink("mysql_default1");
+        StreamSink sink = sinkService.get(mysqlSinkId);
         Assertions.assertEquals(globalGroupId, sink.getInlongGroupId());
-        deleteSink(sinkId);
+        deleteSink(mysqlSinkId);
     }
 
     @Test
     public void testGetAndUpdate() {
-        Integer sinkId = this.saveSink("default_hdfs");
+        Integer sinkId = this.saveSink("mysql_default2");
         StreamSink streamSink = sinkService.get(sinkId);
         Assertions.assertEquals(globalGroupId, streamSink.getInlongGroupId());
 
-        HDFSSink sink = (HDFSSink) streamSink;
+        MySQLSink sink = (MySQLSink) streamSink;
         sink.setEnableCreateResource(InlongConstants.ENABLE_CREATE_RESOURCE);
-        HDFSSinkRequest request = CommonBeanUtils.copyProperties(sink, HDFSSinkRequest::new);
+        MySQLSinkRequest request = CommonBeanUtils.copyProperties(sink, MySQLSinkRequest::new);
         boolean result = sinkService.update(request, globalOperator);
         Assertions.assertTrue(result);
+
         deleteSink(sinkId);
     }
