+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/service/sink/SQLServerSinkServiceTest.java
@@ -15,36 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.core.sink;
+package org.apache.inlong.manager.service.sink;
 
 import org.apache.inlong.manager.common.consts.InlongConstants;
 import org.apache.inlong.manager.common.enums.SinkType;
-import org.apache.inlong.manager.pojo.sink.SinkField;
-import org.apache.inlong.manager.pojo.sink.SinkRequest;
 import org.apache.inlong.manager.pojo.sink.StreamSink;
-import org.apache.inlong.manager.pojo.sink.dlciceberg.DLCIcebergSink;
-import org.apache.inlong.manager.pojo.sink.dlciceberg.DLCIcebergSinkRequest;
+import org.apache.inlong.manager.pojo.sink.sqlserver.SQLServerSink;
+import org.apache.inlong.manager.pojo.sink.sqlserver.SQLServerSinkRequest;
+import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.service.ServiceBaseTest;
 import org.apache.inlong.manager.service.core.impl.InlongStreamServiceTest;
-import org.apache.inlong.manager.service.sink.StreamSinkService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
- * DLCIceberg sink service test..
+ * SQLServer sink service test
  */
-public class DLCIcebergSinkServiceTest extends ServiceBaseTest {
+public class SQLServerSinkServiceTest extends ServiceBaseTest {
 
-    private static final String fieldName = "dlc_field";
-    private static final String fieldType = "dlc_type";
-    private static final Integer fieldId = 1;
-    private final String globalGroupId = "b_group1";
-    private final String globalStreamId = "stream1_dlciceberg";
-    private final String globalOperator = "admin";
+    private static final String globalGroupId = "b_group1_sqlserver";
+    private static final String globalStreamId = "stream1_sqlserver";
+    private static final String globalOperator = "admin";
 
     @Autowired
     private StreamSinkService sinkService;
@@ -55,54 +47,54 @@ public class DLCIcebergSinkServiceTest extends ServiceBaseTest {
      * Save sink info.
      */
     public Integer saveSink(String sinkName) {
-        streamServiceTest.saveInlongStream(globalGroupId, globalStreamId, globalOperator);
-        DLCIcebergSinkRequest sinkInfo = new DLCIcebergSinkRequest();
+        streamServiceTest.saveInlongStream(globalGroupId, globalStreamId,
+                globalOperator);
+        SQLServerSinkRequest sinkInfo = new SQLServerSinkRequest();
         sinkInfo.setInlongGroupId(globalGroupId);
         sinkInfo.setInlongStreamId(globalStreamId);
-        sinkInfo.setSinkType(SinkType.SINK_DLCICEBERG);
-        sinkInfo.setEnableCreateResource(InlongConstants.DISABLE_CREATE_RESOURCE);
-        SinkField sinkField = new SinkField();
-        sinkField.setFieldName(fieldName);
-        sinkField.setFieldType(fieldType);
-        sinkField.setId(fieldId);
-        List<SinkField> sinkFieldList = new ArrayList<>();
-        sinkFieldList.add(sinkField);
-        sinkInfo.setSinkFieldList(sinkFieldList);
+        sinkInfo.setSinkType(SinkType.SINK_SQLSERVER);
+
+        sinkInfo.setJdbcUrl("jdbc:sqlserver://localhost:5432/sqlserver");
+        sinkInfo.setUsername("sqlserver");
+        sinkInfo.setPassword("inlong");
+        sinkInfo.setTableName("user");
+        sinkInfo.setSchemaName("test");
+        sinkInfo.setPrimaryKey("name,age");
 
         sinkInfo.setSinkName(sinkName);
-        sinkInfo.setId((int) (Math.random() * 100000 + 1));
+        sinkInfo.setEnableCreateResource(InlongConstants.DISABLE_CREATE_RESOURCE);
         return sinkService.save(sinkInfo, globalOperator);
     }
 
-    @Test
-    public void testSaveAndDelete() {
-        Integer sinkId = this.saveSink("default1");
-        Assertions.assertNotNull(sinkId);
+    /**
+     * Delete sink info by sink id.
+     */
+    public void deleteSink(Integer sinkId) {
         boolean result = sinkService.delete(sinkId, globalOperator);
         Assertions.assertTrue(result);
     }
 
     @Test
     public void testListByIdentifier() {
-        Integer sinkId = this.saveSink("default2");
+        Integer sinkId = this.saveSink("sqlserver_default1");
         StreamSink sink = sinkService.get(sinkId);
         Assertions.assertEquals(globalGroupId, sink.getInlongGroupId());
-        sinkService.delete(sinkId, globalOperator);
+        deleteSink(sinkId);
     }
 
     @Test
     public void testGetAndUpdate() {
-        Integer sinkId = this.saveSink("default3");
+        Integer sinkId = this.saveSink("sqlserver_default2");
         StreamSink streamSink = sinkService.get(sinkId);
         Assertions.assertEquals(globalGroupId, streamSink.getInlongGroupId());
 
-        DLCIcebergSink sink = (DLCIcebergSink) streamSink;
-        sink.setEnableCreateResource(InlongConstants.DISABLE_CREATE_RESOURCE);
-        SinkRequest request = sink.genSinkRequest();
+        SQLServerSink sink = (SQLServerSink) streamSink;
+        sink.setEnableCreateResource(InlongConstants.ENABLE_CREATE_RESOURCE);
+        SQLServerSinkRequest request = CommonBeanUtils.copyProperties(sink, SQLServerSinkRequest::new);
         boolean result = sinkService.update(request, globalOperator);
         Assertions.assertTrue(result);
 
-        sinkService.delete(sinkId, globalOperator);
+        deleteSink(sinkId);
     }
 
 }
