+++ inlong-manager/manager-service/src/test/java/org/apache/inlong/manager/service/source/TubeMQSourceServiceTest.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.inlong.manager.service.core.source;
+package org.apache.inlong.manager.service.source;
 
 import org.apache.inlong.manager.common.enums.SourceType;
 import org.apache.inlong.manager.pojo.source.StreamSource;
@@ -25,7 +25,6 @@ import org.apache.inlong.manager.pojo.source.tubemq.TubeMQSourceRequest;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.service.ServiceBaseTest;
 import org.apache.inlong.manager.service.core.impl.InlongStreamServiceTest;
-import org.apache.inlong.manager.service.source.StreamSourceService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
