+++ inlong-manager/manager-test/src/main/resources/h2/apache_inlong_manager.sql
@@ -77,7 +77,7 @@ CREATE TABLE IF NOT EXISTS `inlong_cluster_tag`
     `id`          int(11)      NOT NULL AUTO_INCREMENT COMMENT 'Incremental primary key',
     `cluster_tag` varchar(128) NOT NULL COMMENT 'Cluster tag',
     `ext_params`  text                  DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
-    `description`            varchar(256)          DEFAULT '' COMMENT 'Description of cluster tag',
+    `description` varchar(256)          DEFAULT '' COMMENT 'Description of cluster tag',
     `in_charges`  varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
     `status`      int(4)                DEFAULT '0' COMMENT 'Cluster status',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
@@ -103,7 +103,7 @@ CREATE TABLE IF NOT EXISTS `inlong_cluster`
     `ext_tag`      varchar(128)          DEFAULT NULL COMMENT 'Extension tag, for extended use',
     `token`        varchar(512)          DEFAULT NULL COMMENT 'Cluster token',
     `ext_params`   text                  DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
-    `description`            varchar(256)          DEFAULT '' COMMENT 'Description of cluster',
+    `description`  varchar(256)          DEFAULT '' COMMENT 'Description of cluster',
     `heartbeat`    text                  DEFAULT NULL COMMENT 'Cluster heartbeat info',
     `in_charges`   varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
     `status`       int(4)                DEFAULT '0' COMMENT 'Cluster status',
@@ -128,7 +128,7 @@ CREATE TABLE IF NOT EXISTS `inlong_cluster_node`
     `ip`          varchar(512) NOT NULL COMMENT 'Cluster IP, separated by commas, such as: 127.0.0.1:8080,host2:8081',
     `port`        int(6)       NULL COMMENT 'Cluster port',
     `ext_params`  text                  DEFAULT NULL COMMENT 'Another fields will be saved as JSON string',
-    `description`            varchar(256)          DEFAULT '' COMMENT 'Description of cluster node',
+    `description` varchar(256)          DEFAULT '' COMMENT 'Description of cluster node',
     `status`      int(4)                DEFAULT '0' COMMENT 'Cluster status',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
     `creator`     varchar(64)  NOT NULL COMMENT 'Creator name',
@@ -152,6 +152,7 @@ CREATE TABLE IF NOT EXISTS `data_node`
     `username`    varchar(128)          DEFAULT NULL COMMENT 'Username for node if needed',
     `token`       varchar(512)          DEFAULT NULL COMMENT 'Node token',
     `ext_params`  text                  DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
+    `description` varchar(256)          DEFAULT '' COMMENT 'Description of data node',
     `in_charges`  varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
     `status`      int(4)                DEFAULT '0' COMMENT 'Node status',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
