+++ inlong-manager/manager-web/sql/apache_inlong_manager.sql
@@ -160,6 +160,7 @@ CREATE TABLE IF NOT EXISTS `data_node`
     `username`    varchar(128)          DEFAULT NULL COMMENT 'Username for node if needed',
     `token`       varchar(512)          DEFAULT NULL COMMENT 'Node token',
     `ext_params`  text                  DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
+    `description` varchar(256)          DEFAULT '' COMMENT 'Description of cluster tag',
     `in_charges`  varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
     `status`      int(4)                DEFAULT '0' COMMENT 'Node status',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
@@ -596,7 +597,7 @@ CREATE TABLE IF NOT EXISTS `role`
     `id`          int(11)      NOT NULL AUTO_INCREMENT,
     `role_code`   varchar(100) NOT NULL COMMENT 'Role code',
     `role_name`   varchar(256) NOT NULL COMMENT 'Role Chinese name',
-    `disabled`    tinyint(1)  NOT NULL DEFAULT '0' COMMENT 'Whether to disabled, 0: enabled, 1: disabled',
+    `disabled`    tinyint(1)   NOT NULL DEFAULT '0' COMMENT 'Whether to disabled, 0: enabled, 1: disabled',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0 is not deleted, if greater than 0, delete',
     `creator`     varchar(256) NOT NULL COMMENT 'Creator name',
     `modifier`    varchar(256)          DEFAULT NULL COMMENT 'Modifier name',
@@ -617,7 +618,7 @@ CREATE TABLE IF NOT EXISTS `user_role`
     `id`          int(11)      NOT NULL AUTO_INCREMENT,
     `user_name`   varchar(256) NOT NULL COMMENT 'Username',
     `role_code`   varchar(256) NOT NULL COMMENT 'User role code',
-    `disabled`    tinyint(1)  NOT NULL DEFAULT '0' COMMENT 'Whether to disabled, 0: enabled, 1: disabled',
+    `disabled`    tinyint(1)   NOT NULL DEFAULT '0' COMMENT 'Whether to disabled, 0: enabled, 1: disabled',
     `is_deleted`  int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0 is not deleted, if greater than 0, delete',
     `creator`     varchar(256) NOT NULL COMMENT 'Creator name',
     `modifier`    varchar(256)          DEFAULT NULL COMMENT 'Modifier name',
